requirements.txt.txt
nginx


streamlit
pillow